+++ README.md
@@ -5,12 +5,17 @@ This repo offers official implementation for [On the Power of Deep but Naive Par
 ### Reproduction
 For reproducing results for Yahoo! dataset,
 ```
-python3 exps_real.py --dset yahoo --model medium
+python dnpl.py -dset yahoo -model medium
 ```
 
 ### Datasets
 
 All datasets used in this work are provided by [Min-Ling ZHANG](http://palm.seu.edu.cn/zhangml/).
+
+Lost: http://palm.seu.edu.cn/zhangml/files/lost.rar
+MSRCv2: http://palm.seu.edu.cn/zhangml/files/MSRCv2.rar
+Soccer: http://palm.seu.edu.cn/zhangml/files/Soccer%20Player.rar
+Yahoo! News: http://palm.seu.edu.cn/zhangml/files/Yahoo!%20News.rar
 
 ### Reference
 ```
